cuda and opencl plotting examples
